Change Core Android module to Kotlin Module and Move Apollo Schema to new module
