+++ app/src/main/java/com/example/podcastapp/App.kt
@@ -1,7 +1,6 @@
 package com.example.podcastapp
 
 import android.app.Application
-import com.example.core.di.coreModules
 import com.example.data.di.dataModules
 import com.example.podcast_details_ui.di.podcastDetailsModules
 import org.koin.android.ext.koin.androidContext
@@ -20,7 +19,6 @@ class App : Application() {
             androidContext(this@App)
             // Load modules
             modules(
-                coreModules,
                 dataModules,
                 podcastDetailsModules
             )
