+++ app/src/main/java/com/example/podcastapp/MainActivity.kt
@@ -3,39 +3,50 @@ package com.example.podcastapp
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.material3.Text
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
 import com.example.core_ui.theme.PodcastAppTheme
 import com.example.podcast_details_ui.screens.podcastDetails.PodcastDetailsScreen
-import com.google.accompanist.navigation.material.ExperimentalMaterialNavigationApi
-import com.google.accompanist.navigation.material.ModalBottomSheetLayout
-import com.google.accompanist.navigation.material.bottomSheet
-import com.google.accompanist.navigation.material.rememberBottomSheetNavigator
 
 class MainActivity : ComponentActivity() {
 
-    @OptIn(ExperimentalMaterialNavigationApi::class)
+    @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             PodcastAppTheme {
-                val bottomSheetNavigator = rememberBottomSheetNavigator()
-                val navController = rememberNavController(bottomSheetNavigator)
-                ModalBottomSheetLayout(bottomSheetNavigator) {
-                    NavHost(navController, "home") {
-                        composable(route = "home") {
-                            PodcastDetailsScreen(
-                                openPodcastPlayer = {
-                                    navController.navigate("sheet")
-                                }
+
+                var showBottomBar by remember {
+                    mutableStateOf(false)
+                }
+
+                Scaffold(
+                    bottomBar = {
+                        if (showBottomBar) {
+                            Box(
+                                modifier = Modifier
+                                    .fillMaxSize()
+                                    .background(MaterialTheme.colorScheme.primary)
                             )
                         }
-                        bottomSheet(route = "sheet") {
-                            Text("This is a cool bottom sheet!")
-                        }
                     }
+                ) {
+                    PodcastDetailsScreen(
+                        modifier = Modifier.padding(it),
+                        openPodcastPlayer = {
+                            showBottomBar = !showBottomBar
+                        }
+                    )
                 }
             }
         }
