--- core/src/main/java/com/example/core/di/CoreModules.kt
@@ -1,9 +1,0 @@
-package com.example.core.di
-
-import org.koin.dsl.module
-
-val coreModules = module {
-    includes(
-        apiModule
-    )
-}
