+++ data/src/main/java/com/example/data/di/DataModules.kt
@@ -5,6 +5,7 @@ import org.koin.dsl.module
 val dataModules = module {
     includes(
         dataSourceModule,
-        repositoryModule
+        repositoryModule,
+        apiModule
     )
 }
