+++ podcast-details-domain/build.gradle
@@ -10,7 +10,7 @@ java {
 
 dependencies {
     // Modules
-//    implementation project(':core')
+    implementation project(':core')
 
     // Coroutines
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4'
