+++ shared-apollo/src/androidTest/java/com/example/shared_apollo/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.core
+package com.example.shared_apollo
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.podcastapp_core.test", appContext.packageName)
+        assertEquals("com.example.shared_apollo.test", appContext.packageName)
     }
 }
