+++ shared-apollo/src/main/AndroidManifest.xml
