+++ shared-apollo/src/main/graphql/com/example/core/schema.sdl
