+++ shared-apollo/src/test/java/com/example/shared_apollo/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.core
+package com.example.shared_apollo
 
 import org.junit.Test
